rotationing cube

+++ src/vector.h
@@ -2,21 +2,21 @@
 #define VECTOR_H
 
 typedef union {
-  double v[3];
+  float v[3];
 
   struct {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
   };
 } Vector3;
 
 typedef union {
-  double v[2];
+  float v[2];
 
   struct {
-    double x;
-    double y;
+    float x;
+    float y;
   };
 } Vector2;
